remove: AccountHeader만 적용, dev 브랜치에서 Header 컴포넌트 병합 예정
